added jwt authentication and route protection to certain routes

+++ models/event.go
@@ -14,7 +14,7 @@ type Event struct {
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserId      int
+	UserId      int64
 }
 
 func (evt *Event) Save() error {
